pyproject.toml
[project]
name = "mtp"
version = "0.1.0"
description = "Fair Reranker"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ir-datasets>=0.5.11",
    "langchain>=1.2.10",
    "langchain-core>=1.2.14",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.8",
    "pyautocorpus>=0.1.12",
    "python-dotenv>=1.2.1",
    "vllm>=0.16.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]

